Don't check params if they're None
In some cases, especially when the user has not interacted with the GUI
element yet it looks like the parameter params can be None, which leads
to it not being converted to a HotshotXLParams instance. In these cases
we can be pretty sure that the module is disabled, so we handle it the
same way as if params.enable is False.

This does not seem to interrupt image generation, but prints a
stacktrace on the console each time an image is generated.